Converting to 32F, also tried 64F - is latter better on 64-bit machine?  Looks like framerate goes down a bit, now slower than webcam.
git-svn-id: http://binarymillenium.googlecode.com/svn/trunk@691 3bc78312-1b2a-0410-bf3d-39cd8132870d

+++ camthing.cpp
@@ -70,7 +70,7 @@ class CamThing
   {
     count = 0;
 
-    graph = cv::Mat(cv::Size(1280, 720), CV_8UC3);
+    graph = cv::Mat(cv::Size(1280, 720), CV_32FC3);
     graph = cv::Scalar(0);
 
     ///////////////
